Add support for multi-module projects and com.android.test modules.
The DDMS bridge does not seem to operate when destroyed and re-opened.
As a result, multi-module projects were unable to leverage this plugin
to record logcat data.  This change introduces an AdbBridgeManager
singleton to manage DDMS in a single location and spanning the lifecycle
of the build.  This also improves behavior when running Gradle with
--parallel.

As a result of the above, the mechanism used to start the recording
has transitioned from a task dependency to a "doFirst" closure call.
The finalizing LogCatPrinterTask retains the responsibility for
closing out the recording.

Also, expanded the use of this plugin to com.android.test modules
in addition to application and library modules.